+++ src/labs/lab05/example-viscone-full.ts
@@ -1,0 +1,3 @@
+import { ExampleVisibilityCone } from './example-viscone';
+
+export default new ExampleVisibilityCone(document.getElementById('gameCanvas') as HTMLCanvasElement, true);